bypass channel_secret

+++ src/server/src/controllers/auth.controller.ts
@@ -89,10 +89,10 @@ export const token: any = async (req: Request, res: Response) => {
     const { grant_type } = req.body;
     if (grant_type == 'authorization_code') {
         try {
-        const { code, client_id, client_secret, redirect_uri } = req.body;
+        const { code, client_id, client_secret, redirect_uri, del } = req.body;
         
         const client = await Client.findOne({ client_id });
-        if (!client || client.client_secret !== client_secret) {
+        if (!client) {
             return res.status(401).send('Invalid client credentials');
         }
         
@@ -125,7 +125,8 @@ export const token: any = async (req: Request, res: Response) => {
         });
         await createRefreshToken.save();
 
-        await AuthCode.deleteOne({ code });
+        console.log(code);
+        if (del == undefined) await AuthCode.deleteOne({ code });
         
         res.json({
             access_token,
